feat: 🎸 Global
no

BREAKING CHANGE: 🧨 no

✅ Closes: no

+++ src/modules/admin/services/admin.service.ts
@@ -4,8 +4,6 @@ import { InjectMapper } from "@automapper/nestjs";
 
 import { Admin } from "../entities/admin.entity";
 
-import { CreateAdminDto } from "../dtos/create_admin.dto";
-import { CreateSuperAdminDto } from "../dtos/create_super_admin.dto";
 import { ADMIN_MESSAGES } from "../messages/admin.error";
 import { AuthCommonService } from "src/common/auth/services/auth.service";
 import {
@@ -14,12 +12,17 @@ import {
   GetSuccessWithPaginationResponse,
   UpdatedSuccessResponse,
 } from "src/common/response/success.response";
-import { GetAdminReponseDto } from "../dtos/get_admin_response.dto";
-import { PaginationSearchRequestDto } from "src/common/pagination/dtos";
 import { IAdminService } from "../interfaces/admin_service.interface";
 import { AdminRepository } from "../repositories/admin.repository";
 import { AuthAdminTokenRepository } from "../repositories/authAdminToken.repository";
-import { UpdateAdminRequestDto } from "../dtos/update_admin_request.dto";
+import {
+  ChangePasswordAdminRequestDto,
+  CreateAdminRequestDto,
+  CreateSuperAdminRequestDto,
+  GetAdminResponseDto,
+  SearchAdminsRequestDto,
+  UpdateAdminRequestDto,
+} from "../dtos";
 
 @Injectable()
 export class AdminService implements IAdminService {
@@ -31,7 +34,9 @@ export class AdminService implements IAdminService {
     @InjectMapper() private readonly mapper: Mapper,
   ) {}
 
-  async createUser(payload: CreateAdminDto): Promise<CreateSuccessResponse> {
+  async createUser(
+    payload: CreateAdminRequestDto,
+  ): Promise<CreateSuccessResponse> {
     const isExitUser = await this.adminRepository.findByEmail(payload.email);
 
     if (isExitUser) {
@@ -42,7 +47,7 @@ export class AdminService implements IAdminService {
       payload.password,
     );
 
-    const dataForSave: CreateAdminDto = {
+    const dataForSave: CreateAdminRequestDto = {
       ...payload,
       password: hashPassword,
     };
@@ -57,7 +62,7 @@ export class AdminService implements IAdminService {
   }
 
   async createSuperUser(
-    payload: CreateSuperAdminDto,
+    payload: CreateSuperAdminRequestDto,
   ): Promise<CreateSuccessResponse> {
     const godKey = process.env.CREATE_GOD_USER_KEY;
     const defaultPassword = process.env.PASSWORD_GOD_USER;
@@ -79,7 +84,7 @@ export class AdminService implements IAdminService {
     const hashPassword =
       await this.authCommonService.hashPassword(defaultPassword);
 
-    const dataForSave: CreateAdminDto = {
+    const dataForSave: CreateAdminRequestDto = {
       ...payload,
       password: hashPassword,
     };
@@ -94,17 +99,17 @@ export class AdminService implements IAdminService {
   }
 
   async getAdmins(
-    params: PaginationSearchRequestDto,
-  ): Promise<GetSuccessWithPaginationResponse<GetAdminReponseDto[]>> {
+    params: SearchAdminsRequestDto,
+  ): Promise<GetSuccessWithPaginationResponse<GetAdminResponseDto[]>> {
     const { data, pagination } = await this.adminRepository.findAdmins(params);
 
-    const formatData: GetAdminReponseDto[] = this.mapper.mapArray(
+    const formatData: GetAdminResponseDto[] = this.mapper.mapArray(
       data,
       Admin,
-      GetAdminReponseDto,
+      GetAdminResponseDto,
     );
 
-    return new GetSuccessWithPaginationResponse<GetAdminReponseDto[]>(
+    return new GetSuccessWithPaginationResponse<GetAdminResponseDto[]>(
       formatData,
       pagination,
     );
@@ -125,7 +130,73 @@ export class AdminService implements IAdminService {
     return new UpdatedSuccessResponse();
   }
 
+  async changePassword(
+    id: string,
+    data: ChangePasswordAdminRequestDto,
+  ): Promise<UpdatedSuccessResponse> {
+    const user = await this.adminRepository.findByUserId(id);
+
+    if (!user) {
+      throw new BadRequestException(ADMIN_MESSAGES.ADMIN_NOT_FOUND);
+    }
+
+    const isMatchPassword = await this.authCommonService.comparePassword(
+      data.password,
+      user.password,
+    );
+
+    if (!isMatchPassword) {
+      throw new BadRequestException(ADMIN_MESSAGES.PASSWORD_INCORRECT);
+    }
+
+    const hashPassword = await this.authCommonService.hashPassword(
+      data.password,
+    );
+
+    await this.adminRepository.changePassword(id, hashPassword);
+
+    return new UpdatedSuccessResponse();
+  }
+
+  async enableAdmin(id: string): Promise<UpdatedSuccessResponse> {
+    const user = await this.adminRepository.findByUserId(id);
+
+    if (!user) {
+      throw new BadRequestException(ADMIN_MESSAGES.ADMIN_NOT_FOUND);
+    }
+
+    if (user.is_active) {
+      throw new BadRequestException(ADMIN_MESSAGES.USER_WAS_ENABLED);
+    }
+
+    await this.adminRepository.enableAdmin(id);
+
+    return new UpdatedSuccessResponse();
+  }
+
+  async disableAdmin(id: string): Promise<UpdatedSuccessResponse> {
+    const user = await this.adminRepository.findByUserId(id);
+
+    if (!user) {
+      throw new BadRequestException(ADMIN_MESSAGES.ADMIN_NOT_FOUND);
+    }
+
+    if (!user.is_active) {
+      throw new BadRequestException(ADMIN_MESSAGES.USER_WAS_DISABLED);
+    }
+
+    await this.adminRepository.disableAdmin(id);
+
+    return new UpdatedSuccessResponse();
+  }
+
   async deleteAdmin(id: string): Promise<DeletedSuccessResponse> {
+    const isExitUser = await this.adminRepository.findByUserId(id);
+
+    if (!isExitUser) {
+      throw new BadRequestException(ADMIN_MESSAGES.ADMIN_NOT_FOUND);
+    }
+
     await this.adminRepository.deleteAdmin(id);
 
     return new DeletedSuccessResponse();
